refactor : 로또 최소, 최대 금액 상수 위치 수정으로 인한 로직 수정
