+++ src/main/kotlin/domain/model/Lotto.kt
@@ -10,6 +10,9 @@ class Lotto(val numbers: List<LottoNumber>) {
 
     companion object {
         const val LOTTO_SIZE = 6
+        const val LOTTO_MIN = 1
+        const val LOTTO_MAX = 45
+
         const val INVALID_LOTTO_SIZE = "$ERROR 로또 번호는 ${LOTTO_SIZE}개 입니다."
         const val DUPLICATED_LOTTO_NUMBERS = "$ERROR 로또 번호는 중복될 수 없습니다."
     }
