+++ src/main/kotlin/domain/service/LottoGenerator.kt
@@ -8,8 +8,5 @@ interface LottoGenerator {
 
     companion object {
         const val STANDARD_AMOUNT_UNIT = 1000
-        const val STANDARD_AMOUNT_UNIT_KOREAN = "천원"
-        const val LOTTO_MIN = 1
-        const val LOTTO_MAX = 45
     }
 }
