+++ src/main/kotlin/domain/service/LottoGeneratorImpl.kt
@@ -1,22 +1,20 @@
 package domain.service
 
 import domain.model.Lotto
+import domain.model.Lotto.Companion.LOTTO_MAX
+import domain.model.Lotto.Companion.LOTTO_MIN
 import domain.model.Lotto.Companion.LOTTO_SIZE
 import domain.model.LottoNumber
 import domain.model.PurchasePrice
-import domain.service.LottoGenerator.Companion.LOTTO_MAX
-import domain.service.LottoGenerator.Companion.LOTTO_MIN
 import domain.service.LottoGenerator.Companion.STANDARD_AMOUNT_UNIT
-import domain.service.LottoGenerator.Companion.STANDARD_AMOUNT_UNIT_KOREAN
 import util.ErrorConstants.ERROR
 
 class LottoGeneratorImpl(
-    val lottoPrice: Int = STANDARD_AMOUNT_UNIT,
-    val lottoPriceKorean: String = STANDARD_AMOUNT_UNIT_KOREAN,
+    val standardPrice: Int = STANDARD_AMOUNT_UNIT,
 ) : LottoGenerator {
     override fun generate(money: PurchasePrice): List<Lotto> {
         validateMoney(money.value)
-        val purchaseLottoCount = money.value / lottoPrice
+        val purchaseLottoCount = money.value / standardPrice
         return List<Lotto>(purchaseLottoCount) { makeOneLotto() }
     }
 
@@ -26,7 +24,7 @@ class LottoGeneratorImpl(
     }
 
     private fun validateMoney(money: Int) {
-        require(money > 0) { "$ERROR $lottoPriceKorean 이상 입력해주세요." }
-        require(money % lottoPrice == 0) { "$ERROR $lottoPriceKorean 단위로 입력해주세요." }
+        require(money > 0 && money >= standardPrice) { "$ERROR ${standardPrice}원 이상 입력해주세요." }
+        require(money % standardPrice == 0) { "$ERROR ${standardPrice}원 단위로 입력해주세요." }
     }
 }
