+++ src/test/kotlin/domain/service/LottoGeneratorTest.kt
@@ -10,14 +10,14 @@ class LottoGeneratorTest {
     fun `구입 금액이 0 이하면 예외가 발생한다`() {
         assertThrows<IllegalArgumentException> {
             LottoGeneratorImpl().generate(PurchasePrice(0))
-        }.apply { assertThat(this).hasMessage("[ERROR] 천원 이상 입력해주세요.") }
+        }.apply { assertThat(this).hasMessage("[ERROR] 1000원 이상 입력해주세요.") }
     }
 
     @Test
     fun `구입 금액이 천원 단위가 아니면 예외가 발생한다`() {
         assertThrows<IllegalArgumentException> {
             LottoGeneratorImpl().generate(PurchasePrice(10101010))
-        }.apply { assertThat(this).hasMessage("[ERROR] 천원 단위로 입력해주세요.") }
+        }.apply { assertThat(this).hasMessage("[ERROR] 1000원 단위로 입력해주세요.") }
     }
 
     @Test
